error_analyzer: electron_mass.py: ParseDegree(): Change input format.
The original format of (degrees)'(minutes)''(seconds) was a little too
abstract. The new format changes is (degrees)d(minutes)m(seconds)s.
In order to support the case of a plain number being input (i.e., just
the value for degrees, without a 'd' suffixed), the function had to be
changed slightly.